server/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.3.0
cffi==1.17.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.2
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
